module7-task2 | Деструктуризация шапки
